pyproject.toml
[project]
name = "quickfinance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pyside6-fluent-widgets[full]>=1.8.7",
]
